useDropDownClick hook was implemented
